requirements.txt
streamlit~=1.40.1
google.generativeai
